Add files via upload
Client is now added. Also made one or two small changes on Network and Server

There is some things missing, specially error detection and actions, like if a client or a server shutsdown
the network should be able to delete the client/server from its list. Associate client with other server if the last shutsdown, etc
Also, the network is just fully operable if there are the max number of servers active? because if I say the max is 2 and I just
run one server it doesn't work.

I add a load variable because now the election is based on the number of clients but it is supposed to be based on load later on
I think apart from that the base is made. Then we need to implement the actual streaming thing, call methods from the server
associated with the client, etc